[fix] Some build comparisons with missing peers crash 'rpmdeps'
In build comparison runs that see the entire subpackage set change a
lot, the rpmdeps inspection can crash.  The reason is the inspection
code was assuming that there would at least always be an after peer,
but that's not how it works.  You can have a subpackage only present
in the before build and not the after.  So this patch fixes things up
so the rpmdeps inspection does not make that assumption.

Signed-off-by: David Cantrell <dcantrell@redhat.com>