[lib] Drop use_ignore parameter from foreach_peer_file()
This was an idea I had when I first implemented the per-inspection
ignore lists.  I thought we should have the ability to make some
inspection impenetrable to a per-inspection ignore list.  An
inspection so important that it can never be ignored.  Five
inspections carried this classification.  Over time I've come to the
conclusion that this functionality is not really needed and only adds
confusion.  This patch removes that functionality and every inspection
can now have a per-inspection ignore list.

Fixes: #604

Signed-off-by: David Cantrell <dcantrell@redhat.com>