[ci] Pull CentOS images from quay.io, prepare for CentOS 9 Stream
CentOS images are now published on quay.io, so move over to using that
host as the origin.  Requires login credentials, which are stored as
repository secrets in GitHub and then referenced here as variables.

Also start support for building and testing on CentOS 9 Stream.
There's not enough EPEL-9 yet to make that work, so just leave it as a
work in progress.

Signed-off-by: David Cantrell <dcantrell@redhat.com>