[cmd] Fix get_product_release() for single build jobs
In instances where you were running against a single build and not
performing a comparison, rpminspect would fail to correctly match the
right product release name.  The problem was the code path was only
entered for comparison jobs.  But there was a further problem with the
regexp matching.  Some unnecessary checks were in place that I have
removed and now everything behaves as expected.

This problem was first reported here:
https://gitlab.com/redhat/centos-stream/ci-cd/rpminspect-data-centos/-/issues/1

Signed-off-by: David Cantrell <dcantrell@redhat.com>