[ci] Add Slackware Linux 15.0 to the GitHub Actions collection
Slackware Linux 15.0 was recently released and it's now possible to
build and test rpminspect on that platform.  There are some build
dependencies and test dependencies that have to be built by
osdeps/slackware/post.sh, but that's not unlike other systems.

I updated the have_caps_support check in the test suite since rpm on
Slackware Linux is built without libcap support even though it has
libcap on the system.  I also updated the generic.yaml config file to
honor the /usr/man location which is where Slackware installs man
pages.

Signed-off-by: David Cantrell <dcantrell@redhat.com>