[ci] Use -checkgpg=off with slackpkg initially on Slackware Linux
We need to run update-ca-certificates so that slackpkg can import the
GPG key from https://www.slackware.com, but that can't happen until
enough packages are installed to run update-ca-certificates.  So
update the cache with -checkgpg=off, install those packages with
-checkgpg=off, run update-ca-certificates, then update the cache again
the normal way, then reinstall the packages we just installed with GPG
checking enabled.

Signed-off-by: David Cantrell <dcantrell@redhat.com>