[cmd] Implement the -s/--suppress option in rpminspect
This is a command line option you can use to filter the results
printed out by rpminspect in the 'text' and 'summary' output modes.
'json' and 'xunit' still output everything since those output modes
are intended for processing by other systems.

To only report results at the VERIFY severity or higher, pass "-s
VERIFY" on the command line.  By default you get all results.  If no
inspection has any results at or above the suppression level,
rpminspect will skip it in the output entirely.

This option is only for text output.  The exit code is still the same
and all of the results are still computed anyway.  The intent of this
option is to give users a way to just see what needs attention if they
are iterating over fixes and builds.  Eventually you will get to an
exit code of 0 and no output if using the -s option.

Signed-off-by: David Cantrell <dcantrell@redhat.com>