[fix] Handle explicit shared lib deps with %{_isa} notation
Seen while checking the a52dec package.  If a package carries the
correct explicit subpackage Requires dependency when rpmbuild picked
up an automatic shared library dependency, rpminspect should be ok
with that.  It is, except it was not handling the explicit subpackage
Requires syntax that uses %{_isa}, such as:

    Requires: liba52%{_isa} = %{version}-%{release}

This patch fixes rpmdeps so it handles the above.

Signed-off-by: David Cantrell <dcantrell@redhat.com>