[lib] Add libxdiff directly to the source tree
This libxdiff comes from this git project:
    https://github.com/spotrh/libxdiff

This project is available in Fedora, but is not commonly available in
other distributions.  This libxdiff is also from git and libgit2 and
extracted from those projects and turned in to a shared library.  The
previous libxdiff (version 0.23) is API-incompatible and tends to more
available in other distributions than this git/libgit2 libxdiff.  To
make things easier for myself and users, I am just adding this special
libxdiff in to this source tree.

Signed-off-by: David Cantrell <dcantrell@redhat.com>