[lib] Make annocheck failure reporting severity a config file setting
Add a new config file setting under the 'annocheck' block called
'failure_severity'.  The default is VERIFY which means when annocheck
exits non-zero, rpminspect will report that result as VERIFY and
depending on the reporting threshold of rpminspect that may trigger a
non-zero exit code from rpminspect.  The user may modify the config
file for rpminspect and adjust the failure_severity to any valid
result severity.  The most common use case is to set it to INFO to
both report annocheck failures but not fail rpminspect.

Signed-off-by: David Cantrell <dcantrell@redhat.com>