[fix] Honor explicit Requires deps that use zero-epoch syntax
This is entirely valid, just sort of uncommon when you look at a wide
range of RPMs.  By default, a package epoch is 0.  You don't need to
use it until epoch is > 0.  But you can use it.  rpminspect was
assuming people were following this convention and wanted to be graded
according to it.  However, an explicit dependency like this:

    Requires: name-libs = 0:1.0-47

Is the same as:

    Requires: name-libs = 1.0-47

Allow both to exist and report expected changes accordingly.

Signed-off-by: David Cantrell <dcantrell@redhat.com>