[fix] YAML parsing error for the failure_severity setting
Under the 'annocheck' block in the rpminspect config file, you can
define a failure_severity to control what result code annocheck
failures report through rpminspect.  The format of the annocheck block
in the rpminspect config file has changed a bit to accomodate this
setting, but I wanted to retain backwards compatibility with existing
blocks.  So the 'jobs:' section is optional for grouping the annocheck
job definitions.  I need to be careful how I handle parsing this block
to avoid picking up the string 'failure_severity' as a job name.

Signed-off-by: David Cantrell <dcantrell@redhat.com>