[doc] Large set of Doxygen comments in header files
For the next release I want to make sure the library API is documented
and automatically generated for readthedocs.io.  I can use Doxygen for
that, so I need to start updating the comments.  I am putting the API
comments in the public header files and then where necessary reducing
the comments in the library source.

Signed-off-by: David Cantrell <dcantrell@redhat.com>