[fix] Normalize the KABI path and do not warn on access(3) failures
In the kmidiff inspection, get_kabi_file() tries to get a KABI file
for the specified architecture.  If one is not found, the function
returns NULL.  The first issue is that if it is not found, warn() was
called to report that.  This is unnecessary, so the warn() was
removed.  The second problem was with path normalization, which is now
done with abspath().

This is part of the fix for #651

Signed-off-by: David Cantrell <dcantrell@redhat.com>