[cmd] Match products with dist tags containing periods
This bug affects a lot of packages.  Previously rpminspect lacked
support for dist tags with periods in them.  Let's say you had a dist
tag that looked like this:

    .el8_2.2

rpminspect would try to match the product name using the substring
".2".  The reason for this is the match_product() function was taking
the substring from the rindex() position of '.' in the dist tag, which
assumes there's only ever the leading period.

This patch changes the behavior to increment past any and all leading
periods and using the resulting string for the match test.

Signed-off-by: David Cantrell <dcantrell@redhat.com>