[fix] Correct the reporting of kmod parameter differences
This one was hiding in plain sight, but fortunately reported by a
kernel developer.  The code in librpminspect was correctly determining
the kernel module parameter lists differed between the before and
after build, but the reporting made no sense.  It said the module in
the after build both lost and gained the same parameter.

Digging in to it, I made the messages reported by the kmod inspection
a little more explicit.  Telling you what file the issue was found in
and what it was compared to.  The other problem was in
compare_module_parameters().  It was incorrectly computing a list of
added module parameters.  To get that list, the function should have
first intersected the list of before and after parameters, then
computed the difference between that list and the after parameter
list.  Doing that took away the erroneous 'added' reporting.

Fixes: #678

Signed-off-by: David Cantrell <dcantrell@redhat.com>