[lib] Drop the file count and line count checks in 'patches'
The patches inspection no longer produces VERIFY results based on the
file_count_threshold and line_count_threshold settings in the config
file.  These checks were a carryover from rpminspect's ancestor, but
the core problem is that it is trying to enforce an arbitrary rule
across all packages for no good reason.  Who cares if a patch is
subjectively large if it actually fixes a problem?

The patches inspection still reports malformed patches as a BAD
result, but patch statistics are only reported at the INFO level now.

The test suite and example config file has been updated.  The
file_count_threshold and line_count_threshold settings are no longer
honored in the config file.

Signed-off-by: David Cantrell <dcantrell@redhat.com>