[lib] Default the filesize inspection size_threshold to 'info'
Previously the default was to report file size changes of 20% either
growing or shrinking would be reported at the VERIFY level.  The new
default is to report all changes at the INFO level.

The test cases still run with the setting in the config file at 20% to
ensure that users who want to enable the previous behavior of the
filesize inspection will get correct results.

Signed-off-by: David Cantrell <dcantrell@redhat.com>