[fix] Do not report fallthrough changes as VERIFY in changedfiles
In the changedfiles inspection, the last attempt made is to compare
the checksums.  If they are different and it's not a rebase build,
report the change at the VERIFY level.  This doesn't makes sense, it
should be reported at the INFO level.  The critical reporting checks
are above this fallthrough, so really rpminspect should just be
reporting changes for non-rebased builds for informational purposes
only.

Signed-off-by: David Cantrell <dcantrell@redhat.com>