[fix] Honor all per-inspection ignore lists; match path prefix
This is a large one.  First, a number of inspections were not honored
in the config file if you specified a per-inspection ignore list.
That has been fixed and all applicable inspections that can have a
per-inspection list now have one.  Some memory management fixes are
also present as well as a fix that prevents a reset of config file
settings as subsequent files are read.  Settings are now cummulative.
This may require additional work if we find it's really common for an
end user to need to override a config file block, but by and large the
most common case is appending so that's now how all settings are read
in.

The other change is with ignore path matching.  Previously all
matching was done using glob(3).  This is fine and works well, but it
does not allow a user to specify a single directory to exclude and
have it exclude everything in that directory.  The way glob(3) handles
directories in the path name is explicit.  Basically globbing only
works for path components and not the directory names themselves.
This is in POSIX.  While glibc offers some flexibility around this,
it's not portable.

So taking a hint from how rsync does things, rpminspect will now treat
ignore list entries ending with a slash ('/') as a path prefix for
ignore purposes.  So in your rpminspect.yaml file if you specify:

    ---
    virus:
        ignore:
            - /usr/share/mysql-test/

It will ignore /usr/share/mysql-test and everything in that directory.
Without that, the package maintainer would need to explicitly list
glob(3) patterns for each subdirectory construction in that path
prefix which is tedious to maintain and unnecessary.

The data/generic.yaml template file has been updated to reflect all
inspection blocks that were added to honor a per-inspection ignore
list.

Fixes: #691

Signed-off-by: David Cantrell <dcantrell@redhat.com>