[cmd] Improve product release detection for build comparisons
A problem can happen when you are performing a build comparison and
the after build contains an extra architecture not present in the
before build.  The product release detection block looks at the
package peers and takes the first one to check for the product
release.  The problem is the code expects the first peer to have a
before and after build, but that may not be the case.

This patch shifts to using a for loop which looks at each peer and
stops when it find a before and after peer present.  If the loop ends
without getting a before or after release value, then it reports the
builds do not have enough matching peers.

Signed-off-by: David Cantrell <dcantrell@redhat.com>