[tst] Fix the MultipleProvidersCompareRPMs test cases
For rpmdeps, these test cases were expecting a VERIFY result waivable
by Anyone, but that was actually wrong.  The last commit to fix
rpmdeps actually corrected what rpminspect was doing so that these
test cases should be an OK result that are not waivable.  The reason
is the multiple providers check in rpmdeps can't work when comparing
two single RPMs, so it's effectively a no-op.  The multiple providers
check only works when comparing Koji builds because you have all the
subpackages in that case and rpminspect can examine all of them.

Signed-off-by: David Cantrell <dcantrell@redhat.com>