[fix] Correctly check for forbidden directories in RPM payloads
The check for forbidden directory entries made some assumptions about
the RPM payload.  The addedfiles inspection has been modified to check
each path for any forbidden directory as a subdirectory in the full
path.  If that is found, rpminspect reports it.  So you could have an
entry like /usr/lib/.git/something and rpminspect will walk that
backwards and see /usr/lib/.git as a directory and report that.  This
works even if the RPM payloads lacks an explicit /usr/lib/.git entry.

Signed-off-by: David Cantrell <dcantrell@redhat.com>