[fix] Handle more auto deps in the kernel package correctly
The kernel package has some additional deps that need handling by the
rpmdeps inspection.  Namely:

1) Ignore autodep rules of the form kernel(), modalias(), ksym(), and
kmod().  The kmod inspection handles what we need to check for kernel
modules.

2) Expected deprule changes may have the dependency version string end
with ".%{arch}", "+STRING", or ".%{arch}+STRING".  These cases are
checked for as well.

3) Some autodeps are names that do not map to a specific subpackage,
but are a virtual dependency and explicitly matches the package
version-release string.  In these cases the dependency change (if
there is one) is expected.

Fixes: #735

Signed-off-by: David Cantrell <dcantrell@redhat.com>