[fix] Try FNM_LEADING_DIR matches when patterns end in wildcard
This is a common thing as package maintainers have been adding files
to ignore lists. The glob(3) rules do not automatically match a string
like /usr/libexec/ksamples/* as a directory prefix. Technically the *
there will only match one level within /usr/libexec/ksamples. What
I've done to address this is use fnmatch() with FNM_LEADING_DIR in the
flags when the ignore pattern ends with * or ?. I feel this is easier
than asking package maintainers to continually update this list any
time there's another subdirectory added to a full path to ignore. Plus
it's also easier than coming up with some new syntax like ** (used in
Python and rsync and others) or something like
dir:/usr/libexec/ksamples. We can just try the fnmatch() method here
and go with that since it's what the package maintainer is expecting.

Fixes: #740

Signed-off-by: David Cantrell <dcantrell@redhat.com>