[fix] Correctly pick up the use of %autopatch or %autosetup
I had incorrectly written this to ensure we were reading that macro
from %prep, %build, %install, or %check.  But my else block match any
other macro as a section and then stopped reading.  So while it found
the macro, it didn't think it was in a valid section anymore.  My
mistake.  This patch simplifies the spec file reading in the patches
inspection.  rpminspect can't get too detailed here as spec files are
difficult to parse.

Signed-off-by: David Cantrell <dcantrell@redhat.com>