[lib] Make the kmod inspection report changes as INFO only
While working on fixing up rpminspect results for kernel package
builds, it occurred to me that the kmod inspection (which is derived
from what existed in rpminspect's ancestor) is attempting to enforce
some policies that would ensure the Kernel Module Interface remains
the same, unless it's a rebase comparison.  rpminspect's ancestor
predates the kmidiff tool:

    https://sourceware.org/libabigail/manual/kmidiff.html

kmidiff is run by rpminspect as well, so the kmod inspection can be
used for reporting the changes between builds as information.  For
example, what module dependencies changed, what kernel parameters were
gained or lost, and so on.  None of those items impact the KMI, but
are of note for upgrades to the new kernel.  In general, removed
parameters are ignored when loading a module so the fact that one is
lost is not necessarily a bad thing.  Likewise, gaining or losting PCI
device IDs is not really a bad thing either, just the growth of
supported hardware or the retirement of formerly supported hardware.

Signed-off-by: David Cantrell <dcantrell@redhat.com>