[fix] Simplify severity reporting in the changedfiles inspection
Report file reading errors at the RESULT_BAD level and
security-related path changes at the VERIFY level, but other changes
should be reported as INFO.  Regardless of the comparison being a
rebase or not.  In every instance, changes are deliberate, but we want
the warnings to catch changes to security-related paths.

Signed-off-by: David Cantrell <dcantrell@redhat.com>