[fix] Fail if we cannot read RPMs before downloading
rpminspect tries to determine the amount of storage space required of
all RPMs in a build before downloading it.  It does this by using
libcurl to get the download size of each package by URL.  If there's
an error, we get !CURLE_OK and the size is 0.  Report this, but don't
continue trying to download the packages.

Fixes: #770

Signed-off-by: David Cantrell <dcantrell@redhat.com>