[lib] Move ./rpminspect.yaml reading to init_rpminspect()
This is a small optimization.  The optional local rpminspect.yaml was
read in from the rpminspect(1) code by calling init_rpminspect()
again.  The problem there is it runs through all the initialization
code again.  It doesn't hurt anything, but is unnecessary.  So move
the ./rpminspect.yaml read in to init_rpminspect().  In the
rpminspect(1) put one last failsafe in to try and read in a local
rpminspect.yaml file if no initialization has happened yet.

Signed-off-by: David Cantrell <dcantrell@redhat.com>