[cfg] Clarify the 'ignore' block in comments
Clarify the syntax for entries in ignore blocks in the rpminspect
config files.  The ignore block, either the global one or any that go
under a specific inspection, can have entries that fit two
classifications:

* glob(7)-style syntax.  Matching is performed using fnmatch(3), so
  not all aspects of glob(7) syntax apply because rpminspect is
  matching against strings and not the actual filesystem.

* Path prefixes, which you get by ending the entry with a slash ('/')

Many people have already figured this out, but I wanted the default
config file template to explain it.

Signed-off-by: David Cantrell <dcantrell@redhat.com>