[lib] Support optional product release configuration files
This is a new thing that data packages can provide in addition to the
main configuration file and profiles.  rpminspect determines what
product release string to use for a specific job.  This string maps to
data files in the corresponding data package.  What this patch does is
adds support for the data package to deliver per-product-release
configuration profiles and have them automatically load.

These can now live in:
    /usr/share/rpminspect/VENDOR/product-release/PRODUCT_RELEASE_STRING.yaml

For example.  Say Fedora wanted an fc36 configuration file in
rpminspect-data-fedora.  It could add:

    /usr/share/rpminspect/fedora/product-release/fc36.yaml

If rpminspect is invoked with the fedora.yaml configuration file and
the job is for fc36 packages, it will check for the above fc36.yaml
file and if it's found, load it after it loaded the main configuration
file and optional profile, but before the ./rpminspect.yaml file that
may exist.

The purpose of this change is to allow for rule definitions across
entire product releases without needing to create a separate
rpminspect data package.

Signed-off-by: David Cantrell <dcantrell@redhat.com>