[fix] Support per-file allowed lists for the badfuncs inspection
Some programs and libraries deliberately use functions that have been
deemed forbidden by policy.  An example would be a library providing
runtime support for older software.  In cases like this, you can
specify an allowed block listing path globs and the specific forbidden
functions they are allowed to use.  It is adviseable to add a comment
idnicating why the forbidden function is allowed so future readers
will understand why the rpminspect.yaml file is configured that way.

The path specifications can be explicit paths (local path notation,
relative to /) or patterns that are compatible with fnmatch(3) and
glob(3).

For example:

    ---
    badfuncs:
        allowed:
            /usr/sbin/somenetcmd:
                - inet_aton
                - gethostbyname
            /usr/lib*/libsomething.so.*:
                - inet_network

You can place this block in the package's rpminspect.yaml file to
further control how the badfuncs inspection works for your build.

Fixes: #573

Signed-off-by: David Cantrell <dcantrell@redhat.com>