[fix] Use allowed_arch() in the arch and subpackages inspections
If '-a' is specified on the command line and the program has an
explicit list of architectures, filter out other architectures with a
call to allowed_arch() for the arch and subpackages inspections.  In
some cases it appears that non-desired architectures are included in
these inspections anyway which may be because the program is skipping
the downloading part which is how -a was initially implemented.

Fixes: #807

Signed-off-by: David Cantrell <dcantrell@redhat.com>