[fix] Handle RPM payloads with individual entries >2GB
RPM uses cpio for the payload stream.  cpio has an individual entry
size limit of 2GB.  In cases where the RPM payload carries an entry
larger than that, librpminspect fails to extract the contents of the
RPM because libarchive can't handle this stream.  The workaround is to
use librpm and libarchive to convert the payload stream to a pax
archive and then let librpminspect extract that.  The conversion is
done in a manner similar to the rpm2archive tool.

Fixes: #820

Signed-off-by: David Cantrell <dcantrell@redhat.com>