[lib] Only support RPM payload conversion with newer librpm releases
At least on RHEL-7, the rpm version there lacks the rpmfiles.h and
rpmarchive.h headers which from what I can tell in the rpm source
means that API was not yet public in the release of rpm included with
RHEL-7.  Since this payload conversion is a value-add for rpminspect
for supporting large files in the payload stream, I think it's ok to
restrict it to more recent librpm versions.

Signed-off-by: David Cantrell <dcantrell@redhat.com>