[fix] Report removed patches by checking RPMTAG_PATCHES list
The patches inspection performs a number of checks on patch files in
SRPMs.  The last thing it does is report any removed patches when
comparing two builds.  However, this removal check was done checking
the entire list of files in the SRPM rather than just looking at the
RPMTAG_PATCH files from the SRPM header.  This patch corrects that
error.

Fixes: #802

Signed-off-by: David Cantrell <dcantrell@redhat.com>