[fix] Do not remove all () substrings from requirements in rpmdeps
In the rpmdeps inspection, a check is performed to ensure subpackages
with automatic shared library dependencies carry an explicit named
dependency on the subpackage providing that shared library.  Part of
that comparison trims the ISA substring that appears in the
requirement strings, but that was making the assumption that all
substrings in parens were ISA substrings.  That's not the case.  This
patch takes care to leave any paren substrings in place and only
remove the ISA substrings.

The ISA substrings are matched using a regular expression.  This is
done to avoid listing out all possible ones as defined in
/usr/lib/rpm/platform/*/macros since that may change from time to
time.

Fixes: #799

Signed-off-by: David Cantrell <dcantrell@redhat.com>