[lib] Support a new config file block called 'environment'
In addition to letting users specify the product release string on the
command line, also introduce a block in the config file where you can
do the same thing.  It would look like this:

    environment:
        product_release: el8

Use this in a per-package rpminspect.yaml file where rpminspect is
unable to detect the product release string from the dist tag.  Some
packages lack a dist tag.

Fixes: #819

Signed-off-by: David Cantrell <dcantrell@redhat.com>