[lib] Add headers dir support back to the the abidiff inspection
The previous method was not working, so modify things to pass the
correct paths to abidiff(1).  Now the inspection passes debuginfo
paths and header paths.

There may be some issues regarding private headers, but I'm not aware
of any package out there that ships private headers that are installed
on the system.  But I could be wrong.

Signed-off-by: David Cantrell <dcantrell@redhat.com>