[lib] When downloading modules, only display Downloading msg once
This is a side effect of the way rpminspect stores the module
information internally.  Regular Koji builds have one buildentry, but
modules have any number of them.  The download loop iterates over the
build entries and for each one displays the "Downloading..." message.
For regular Koji builds this just appears once, but for modules it
appears for each buildentry.  Modify the loop so it only displays once
for module downloads.

Signed-off-by: David Cantrell <dcantrell@redhat.com>