[fix] Only report multiple subpackages Provides for explicit Requires
In the rpmdeps inspection, only report multiple subpackages Provide
the same thing when there are one or more explicit Requires for one of
the providing subpackages.  The recent example that this affects is
the curl package in Fedora.  curl now provides both libcurl and
libcurl-minimal.  Either library subpackage can be installed to
satisfy the requirement of /usr/bin/curl, so both library subpackages
provide libcurl.os.N.  However, since the user has a runtime choice,
no subpackage should carry an explicit Requires on 'libcurl' or
'libcurl-minimal'.

Signed-off-by: David Cantrell <dcantrell@redhat.com>