[fix] Only report missing explicit Requires for single providers
In cases where a package gains an automatic ELF Provides for a shared
library, subpackages that depend on that library need to also carry an
explicit Requires with the name and version of the subpackage carrying
the shared library.  Only report these as missing where there is only
a single subpackage providing the shared library.  In the case of
multiple providers, you don't want any subpackage to carry an explicit
Requires.

Signed-off-by: David Cantrell <dcantrell@redhat.com>