[fix] Check for completed builds and closed tasks when fetching
When gathering a Koji build, stop if the build state is not
BUILD_COMPLETED.  Likewise, when gathering a Koji task, stop if the
task state is not TASK_CLOSED.  Any other state indicates a failed
condition that either happened by itself or the user triggered.  In
any of those cases, we do not want to try to fetch those artifacts.

Fixes: #881

Signed-off-by: David Cantrell <dcantrell@redhat.com>