[cmd] Add the -b/--build-type command line option to rpminspect
This command line option lets the user set the Koji build type of the
inputs given.  Normally rpminspect will automatically detect the build
type, but sometimes that may fail or you find yourself in a situation
where you need to tell rpminspect the type.  The -b option is for
that.  List available build types with the -l command line option.

Signed-off-by: David Cantrell <dcantrell@redhat.com>