[lib] Refactor code to make better use of list_add()
Refactor a lot of code to make better use of list_add() where
possible.  This is just a function I wrote to handle adding a new
'char *' to a string_list_t.  No reason not to use it, but I think
when I wrote list_add() it was after all of the other stuff had been
written.

Signed-off-by: David Cantrell <dcantrell@redhat.com>