[cmd] Trim the leading period from the product_release string
The product release string is based on the concept of "dist tags" seen
in Fedora-derived distributions.  When matching it, rpminspect starts
with the RPMTAG_RELEASE value and captures the part of the string
starting with a period since this is typically where the dist tag
begins.  However, it should not save the period as
ri->product_release.  The product_release value is used throughout
librpminspect when determining per-product rules from the vendor data.

Signed-off-by: David Cantrell <dcantrell@redhat.com>