[lib] Get list of %*auto* macros from the config file in 'patches'
The patches inspection looks for any defined Patch files that do not
have a corresponding %patch macro.  It reports any missing a %patch
macro as an error, unless it finds an auto patching macro.  Until this
patch, that was just %autosetup or %autopatch.  But since auto
patching macros can be defined at runtime in rpm macro files, move the
list of auto patching macros to check for to the config file.

The config file section now adds an "automacros" list under the
"patches" section.  Like this:

    patches:
        # List of spec file macros that handle automatic patching.  If
        # one of these is found, rpminspect will not report defined
        # patches missing corresponding %patch macros.  NOTE: Do not
        # list the macros here with '%' prefixes or braces ('{' or
        # '}').  Just list their names.
        automacros:
            - autopatch
            - autosetup
            - forgeautosetup

Now, when a new auto patching macro is created, only the vendor data
package and rpminspect config file need to change rather than the
rpminspect code.

Fixes: #921

Signed-off-by: David Cantrell <dcantrell@redhat.com>