[lib] Change how RPMs are unpacked when peering happens
Previously rpminspect would unpack each RPM in to its own
subdirectory.  Lots of inspections need to examine the files inside an
RPM, so this seemed like a reasonable way to do that.  However, some
inspections need a view of the entire built package installed.  Things
like abidiff want to see the debuginfo and libraries and headers.
Depending on what the packager did, those files may all be in one
package or multiple subpackages.  To make this easier for inspections
that need to see all files extracted, rpminspect now extracts all RPMs
for each build to the same root tree (technically one per architecture
with src and noarch separated out).  rpminspect still has all of the
RPM header information and knows what subpackage a file belongs to, so
this is just an on-disk representation to make things easier for
different inspections.

Signed-off-by: David Cantrell <dcantrell@redhat.com>