[tst] Sync up Copyright lines with format used in lib and src
Strike the copyright symbol (a circled capital letter C) and just use
the word Copyright.  The symbol is Unicode character U+00A9 and
occassionally I would hear from people asking why there's a question
mark after Copyright because they would be using a non-Unicode capable
text editor or some other environment improperly configured for
Unicode.  Because of this, I removed the symbol from the lib and src
files, but forgot test.  Copyright statements do not need to actually
use the symbol.  You can use the symbol, "Copyright", or "Copr.".
This is from US Copyright Law.  While the enclosed C is still used all
over the place, it is not generally required under the Berne
Convention.  Also, ownership of a work is implied unless explicitly
disclaimed anyway and lawyers have all sorts of fun figured out
ownership of things.

SO...to make the copyright notices in this project simple, the format
of "Copyright YEAR" will be used.  Note that "year ranges" are not
really a thing either.  You set the year and then that's it.  Do not
use year ranges in these files.

Signed-off-by: David Cantrell <dcantrell@redhat.com>