[lib] Modify the result reporting functions so waiverauth is optional
waiverauth is technically optional.  I mean, it shouldn't even exist
in the first place, but that's a discussion for another day.  In cases
where an inspection is skipped or is otherwise OK, there's actually
nothing to waive.  Not Waivable is not wrong, but also not correct.
In these cases I just don't want it to show anything.  So modify the
results reporting function to not assume waiverauth is actually set to
anything.

Signed-off-by: David Cantrell <dcantrell@redhat.com>